Add support for backup filetype detection
The standard library sndhdr is deprecated and will be removed in Python
3.13: https://peps.python.org/pep-0594/#sndhdr

This change adds support for detecting filetypes with the
  `filetype` module (https://github.com/h2non/filetype.py) instead.

That means unfortunately `wav2vec` no longer has zero dependencies for
Python 3.13+ users, but `filetype` is a small, pure-python (portable) library.

This change also improves error reporting in the case that the input
file type cannot be detected.